Add function to create key pair using a seed
Seeding may be useful to enhance entropy, if the main entropy source is not perfect, and if the additional entropy source is not too bad. Even if the main entropy source was perfect, it should not hurt to use a less-than-perfect, but also not-bad, additional entropy source.
SHA1PRNG is used because the default RNG varies between platforms and might simply ignore the seed. However, the SHA1PRNG is always the same and always secure enough.